Revised Orion ATA driver with ext2 and fat updates.
The updated ATA driver accepts a channel number as the region
LOCATION.  This change trickes into the filesystem drivers with a new
syntax for combining regions.

The describe_size function is factored out of the image code and is
usable throughout the loader.

The region parser has been fixed so that START offsets and PATHs work
well together.

The filesystem drivers use a CRC to determine when the relevent
block_driver environment variable is changed.